feat: renamed test-helper package and refactored utility functions
