+++ packages/test-utilities/jest.config.js
@@ -3,5 +3,6 @@ const { jest: lernaAliases } = require('lerna-alias');
 module.exports = {
   preset: 'ts-jest',
   moduleNameMapper: lernaAliases(),
+  setupFilesAfterEnv: ["jest-extended/all"],
   coverageReporters: [['lcov', { projectRoot: '../../' }]]
 };
