+++ packages/test-utilities/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@apimatic/test-helper",
+  "name": "@apimatic/test-utilities",
   "version": "0.0.0",
   "description": "provides the assertion utilities",
   "main": "lib/index.js",
@@ -41,6 +41,7 @@
     "@size-limit/preset-small-lib": "^7.0.8",
     "babel-plugin-annotate-pure-calls": "^0.4.0",
     "jest": "^26.4.2",
+    "jest-extended": "^4.0.2",
     "jsdom": "^19.0.0",
     "jsdom-global": "^3.0.2",
     "lerna-alias": "3.0.3-0",
@@ -61,6 +62,6 @@
   "repository": {
     "type": "git",
     "url": "git@github.com:apimatic/apimatic-js-runtime.git",
-    "directory": "packages/test-helper"
+    "directory": "packages/test-utilities"
   }
 }
