+++ packages/test-utilities/src/index.ts
@@ -1,0 +1,2 @@
+export * from './streamUtils';
+export * from './assertionUtils';
