+++ tsconfig.monorepo.json
@@ -35,7 +35,7 @@
       "path": "./packages/oauth-adapters"
     },
     {
-      "path": "./packages/test-helper"
+      "path": "./packages/test-utilities"
     }
   ]
 }
